requirements.txt
python-binance==1.0.19
python-dotenv==1.0.1
typer==0.12.3
click==8.1.7
rich==13.7.1
requests==2.31.0
